refactor(tools): separate time functions from tools.utils
KAG-2955
