+++ kong-3.6.0-0.rockspec
@@ -172,6 +172,7 @@ build = {
     ["kong.tools.uuid"] = "kong/tools/uuid.lua",
     ["kong.tools.rand"] = "kong/tools/rand.lua",
     ["kong.tools.system"] = "kong/tools/system.lua",
+    ["kong.tools.time"] = "kong/tools/time.lua",
 
     ["kong.runloop.handler"] = "kong/runloop/handler.lua",
     ["kong.runloop.events"] = "kong/runloop/events.lua",
